pyproject.toml
[project]
name = "corpuscraft"
version = "0.1.0"
description = "Transform your documents into training datasets with local-first synthetic data generation"
readme = "README.md"
authors = [
    { name = "Belg PC", email = "tomsomoza911@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["dataset", "synthetic-data", "qa", "embeddings", "llm", "docling", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
]

dependencies = [
    # Document parsing
    "docling>=2.60.1",

    # LLM integration
    "ollama>=0.3.0",

    # CLI and config
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",

    # Data handling
    "datasets>=2.14.0",

    # Text processing
    "langchain-text-splitters>=0.2.0",
]

[project.optional-dependencies]
# Additional LLM backends
transformers = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "accelerate>=0.20.0",
]
vllm = [
    "vllm>=0.4.0",
]
cloud = [
    "openai>=1.0.0",
    "anthropic>=0.25.0",
]

# Quality and validation
quality = [
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
]

# Web UI
ui = [
    "gradio>=4.0.0",
]

# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

# All optional dependencies
all = [
    "corpuscraft[transformers,vllm,cloud,quality,ui]",
]

[project.scripts]
corpuscraft = "corpuscraft.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/corpuscraft"
Repository = "https://github.com/yourusername/corpuscraft"
Issues = "https://github.com/yourusername/corpuscraft/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=corpuscraft --cov-report=term-missing"
